telegram_bot/requirements.txt
requests
pymongo
feedparser
ipdb
coloredlogs
python-telegram-bot
